# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Học Tập",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/setup/status": {
      "get": {
        "tags": [
          "setup"
        ],
        "summary": "Get Setup Status",
        "operationId": "get_setup_status",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetupStatus"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/setup": {
      "post": {
        "tags": [
          "setup"
        ],
        "summary": "Complete Setup",
        "operationId": "complete_setup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Setup already done",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "Profile": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "avatar": {
            "type": "string",
            "enum": [
              "cat",
              "dog",
              "rabbit",
              "bear",
              "fox",
              "panda"
            ],
            "title": "Avatar"
          },
          "grade": {
            "type": "integer",
            "title": "Grade"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "avatar",
          "grade"
        ],
        "title": "Profile"
      },
      "ProfileIn": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "avatar": {
            "type": "string",
            "enum": [
              "cat",
              "dog",
              "rabbit",
              "bear",
              "fox",
              "panda"
            ],
            "title": "Avatar"
          },
          "grade": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Grade"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "avatar",
          "grade"
        ],
        "title": "ProfileIn"
      },
...
      "SetupRequest": {
        "properties": {
          "pin": {
            "type": "string",
            "pattern": "^[0-9]{4}$",
            "title": "Pin"
          },
          "pin_confirm": {
            "type": "string",
            "pattern": "^[0-9]{4}$",
            "title": "Pin Confirm"
          },
          "profile": {
            "$ref": "#/components/schemas/ProfileIn"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "pin",
          "pin_confirm",
          "profile"
        ],
        "title": "SetupRequest"
      },
      "SetupStatus": {
        "properties": {
          "setup_required": {
            "type": "boolean",
            "title": "Setup Required"
          }
        },
        "type": "object",
        "required": [
          "setup_required"
        ],
        "title": "SetupStatus"
      },
...
    }
  }
}